requirements.txt
numpy==1.18.2
psycopg2-binary==2.8.4